refactor: make menu component
* stop stuffing everything in scoreboard
