+++ src/components/Menu/index.scss
@@ -1,0 +1,16 @@
+.game-menu {
+    max-width: 1000px;
+
+    @media (max-width: 600px) {
+        // less padding for smaller screens
+        padding: 8px 8px;
+
+        .MuiDialogContent-root {
+            padding: 8px 8px;
+        }
+    }
+
+    .MuiDialogContent-root {
+        height: 75vh;
+    }
+}
